add kms/pgp bootstrapping commands for signing
This adds the "cmrel bootstrap-pgp" command which will start a
cloudbuild job to both bootstrap a PGP identity and also return the raw
PEM-encoded public key.

go.mod is also updated to include the golang crypto libraries, which are
required by the code we copy-paste from the heptio/google-kms-pgp
project

We can't import google-kms-pgp directly because it's outdated and
doesn't contain the support for SHA512 which we need. We also modify it
slightly to remove the need to query a key version to get metadata; this
requires additional permissions which we don't necessarily want to grant
to our signing cloudbuild jobs.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>