re-tag images when publishing, don't use stored image name
The makefile flow tags images with names such as
"cert-manager-acmesolver-amd64:v1.8.0-alpha.2", which do not include the
name of the registry to which the image should be pushed. That's
intentional; we don't want or need to tie the name of the registry or
registries we might push to at build time.

In other words, for a given release X we don't want to have to release
X+1 in order to push images to a different registry.

This change re-tags images after they're loaded but before they're
pushed, using the value of the registry name we previously used for
validation. It also removes that validation, since it wasn't
particularly useful.

Also makes a few changes which make this task easier.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>