use make upload-release to create a release
A goal for the make-based build system is that all commands to build and
stage a release are defined in the cert-manager repo, making the release
repo sugar around the core of the cert-manager repo. This should make it
much easier for new developers to use.

This commit uses the upload-release target which was added to
cert-manager in the following PRs:

release-1.8: https://github.com/cert-manager/cert-manager/pull/5209
     master: https://github.com/cert-manager/cert-manager/pull/5205

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>