reduce work + simplify output
The path to which we write staged releases is now simpler for tagged
releases, and the release process has been simplified to reflect this.

We no longer need to calculate this path, nor do we need to output it.
A maintainer performing a release can simply use their RELEASE_VERSION
variable which they already have.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>