print to stdout with unknown output format
Before this change, an unknown value for the output format option
would've led to cmrel silently exiting and printing / doing nothing.

In preference to erroring out, this commit makes the default action
be to print to stdout.

Also checks a lowercased version of the argument so that FILE will
still print to a file.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>